local update
