--- agent/src/theory_expert_agent/__init__.py
@@ -1,6 +1,0 @@
-# Theory Expert Agent Package 
-# 创建全局实例
-from .agent import LightRAGAgent
-
-lightrag_agent = LightRAGAgent()
-lightrag_agent.initialize()
