+++ agent/src/utils/lightrag_tool/__init__.py
@@ -1,0 +1,8 @@
+# Theory Expert Agent Package 
+# 创建全局实例
+from .LightRAGTool import LightRAGTool
+
+lightragTool = LightRAGTool()
+lightragTool.initialize()
+lightrag_tools = lightragTool.create_tools()
+
